Create Note Component for displaying inline
So we moved a a function component using useEffect and removing the variables you aren't using at the moment.

I think storing the note as a data attribute on the article itself might make more sense but then you would need to be generating components inside interweave which might be difficult.

AlignNotes is outside the component so it's not regenerated on each rerender.